% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liger.R
\name{calcARI}
\alias{calcARI}
\title{Calculate adjusted Rand index}
\usage{
calcARI(object, clusters.compare)
}
\arguments{
\item{object}{\code{liger} object. Should run quantileAlignSNF before calling.}

\item{clusters.compare}{Clustering with which to compare (named vector).}
}
\value{
Value of ARI

Adjusted Rand index value.
}
\description{
Computes adjusted Rand index for \code{liger} clustering and external clustering.
The Rand index ranges from 0 to 1, with 0 indicating no agreement between clusterings and 1
indicating perfect agreement.
}
