% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liger.R
\name{calcPurity}
\alias{calcPurity}
\title{Calculate purity}
\usage{
calcPurity(object, classes.compare)
}
\arguments{
\item{object}{\code{liger} object. Should run quantileAlignSNF before calling.}

\item{classes.compare}{Clustering with which to compare (named vector).}
}
\value{
Purity value.
}
\description{
Calculates purity for \code{liger} clustering and external clustering (true clusters/classes).
Purity can sometimes be a more useful metric when the clustering to be tested contains more
subgroups or clusters than the true clusters (or classes). Purity also ranges from 0 to 1,
with a score of 1 representing a pure, or accurate, clustering.
}
