% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checkAnnoArguments}
\alias{checkAnnoArguments}
\title{checkAnnoArguments}
\usage{
checkAnnoArguments(argList)
}
\arguments{
\item{argList}{A list of arguments passed into 'runScAnnotation".}
}
\description{
checkAnnoArguments
}
