% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checkCombArguments}
\alias{checkCombArguments}
\title{checkCombArguments}
\usage{
checkCombArguments(argList)
}
\arguments{
\item{argList}{A list of arguments passed into 'runScCombination".}
}
\description{
checkCombArguments
}
