% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checkStatArguments}
\alias{checkStatArguments}
\title{checkStatArguments}
\usage{
checkStatArguments(argList)
}
\arguments{
\item{argList}{A list of arguments passed into 'runScStatistics".}
}
\description{
checkStatArguments
}
