% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAnnotation.R
\name{clusterBarPlot}
\alias{clusterBarPlot}
\title{clusterBarPlot}
\usage{
clusterBarPlot(
  cell.annotation,
  sel.col = "Cell.Type",
  cell.colors = NULL,
  legend.title = NULL,
  legend.position = "bottom",
  legend.ncol = NULL
)
}
\arguments{
\item{cell.annotation}{A data.frame of cells' annotation containing the cells' Cluster and other information to be colored.}

\item{sel.col}{The column name of cell.annotation, which indicating the type of cells.}

\item{cell.colors}{An array of colors used to show the cells' type. If NULL, the default colors will be used.}

\item{legend.title}{The title of legends. If NULL, the value of "sel.col" will be used.}

\item{legend.position}{The position of legends ("none", "left", "right", "bottom", "top", or two-element numeric vector).}

\item{legend.ncol}{The number of column of legends.}
}
\value{
A bar plot.
}
\description{
clusterBarPlot
}
