% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ConfusionMatrix}
\alias{ConfusionMatrix}
\title{Heatmap of classification result.}
\usage{
ConfusionMatrix(label.name, label, predict)
}
\arguments{
\item{label}{A vector of the original labels for each cell in the test set.}

\item{predict}{A vector of the predicted labels for each cell in the test set.}

\item{label.name}{A vector of the sorted unique labels}
}
\value{
A heatmap for the confusion matrix of the classification result.
}
\description{
Heatmap of classification result.
}
