% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtypeFunction.R
\name{Entropy_test}
\alias{Entropy_test}
\title{Entropy_test}
\usage{
Entropy_test(expr, k = 1000)
}
\arguments{
\item{expr}{Rows should be cells and the last column should be "label".}

\item{k}{Number of genes to select.}
}
\value{
'out'  selected gene list(class "character)
}
\description{
Entropy_test
}
