% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extractFiles}
\alias{extractFiles}
\title{extractFiles}
\usage{
extractFiles(
  savePath,
  sampleNames,
  outputPath,
  files = c("report-scStat.html", "report-scAnno.html"),
  subfolders = NULL
)
}
\arguments{
\item{savePath}{A path of samples' result folder.}

\item{sampleNames}{A vector of samples' names (the subfolder names in 'savePath').}

\item{outputPath}{A path to saving the extracted reports.}

\item{files}{The name of files you want to extract. The default is c("report-scStat.html", "report-scAnno.html").}

\item{subfolders}{The name of subfolders for the files you want to extract. The default is NULL.
It can be a character string, which means all files are under the subfolder.
It can also be a character string vector with same length as "files", which are corresponding to "files".}
}
\description{
Extract files from each sample's folder and rename them with sample's name.
}
