% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAnnotation.R
\name{genAnnoReport}
\alias{genAnnoReport}
\title{genAnnoReport}
\usage{
genAnnoReport(results, savePath)
}
\arguments{
\item{results}{A list generated by 'runScAnnotation'}

\item{savePath}{A path to save the results files. If NULL, the 'statPath' will be used instead.}
}
\description{
genAnnoReport
}
