% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate10Xdata}
\alias{generate10Xdata}
\title{generate10Xdata}
\usage{
generate10Xdata(matrix, gene.info, outPath, overwrite = F)
}
\arguments{
\item{matrix}{A gene-cell matrix or data.frame.}

\item{gene.info}{A data.frame of gene information. It should contain two columns,
the first is gene Ensemble ID, and the second is gene symbol.
The order of the genes should be consistant with the row order of 'matrix'.}

\item{outPath}{A path to save the output files.}

\item{overwrite}{If TRUE and the output file already exists, the file is
silently overwritten, otherwise an exception is thrown. The default is "FALSE".}
}
\description{
Generate a 10X-like data folder based on the data matrix and gene information,
which can be used directly to perform scCancer analysis.
}
