% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scStatistics.R
\name{genStatReport}
\alias{genStatReport}
\title{genStatReport}
\usage{
genStatReport(results, savePath)
}
\arguments{
\item{results}{A list generated by 'runScStatistics'}

\item{savePath}{A path to save the results files(suggest to create a foler named by sample name).}
}
\description{
genStatReport
}
