% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getCellTypeColor}
\alias{getCellTypeColor}
\title{getCellTypeColor}
\usage{
getCellTypeColor(cell.types)
}
\arguments{
\item{cell.types}{A vector of cell types.}
}
\value{
A vector of colors.
}
\description{
getCellTypeColor
}
