% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getDefaultColors}
\alias{getDefaultColors}
\title{getDefaultColors}
\usage{
getDefaultColors(n = NULL, type = 1)
}
\arguments{
\item{n}{The number of colors.}

\item{type}{The type of color style. Only 1, 2, or 3 is allowed.}
}
\value{
A vector of colors.
}
\description{
getDefaultColors
}
