% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAnnotation.R
\name{getDefaultGeneSets}
\alias{getDefaultGeneSets}
\title{getDefaultGeneSets}
\usage{
getDefaultGeneSets(species = "human")
}
\arguments{
\item{species}{A character string indicating what species the sample belong to.
Only "human"(default) or "mouse" are allowed.}
}
\value{
A list of gene sets (50 hallmark gene sets).
}
\description{
getDefaultGeneSets
}
