% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getDefaultMarkers}
\alias{getDefaultMarkers}
\title{getDefaultMarkers}
\usage{
getDefaultMarkers(species = "human")
}
\arguments{
\item{species}{A character string indicating what species the sample belong to.
Only "human"(default) or "mouse" are allowed.}
}
\value{
A list of default markers of several common cell types.
}
\description{
Return default markers of several common cell types.
}
