% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAnnotation.R
\name{getFilterData}
\alias{getFilterData}
\title{getFilterData}
\usage{
getFilterData(dataPath, statPath, savePath = NULL,
  bool.filter.cell = T, bool.filter.gene = T,
  anno.filter = c("mitochondrial", "ribosome", "dissociation"),
  nCell.min = 3, bgPercent.max = 0.1, hg.mm.mix = F)
}
\arguments{
\item{dataPath}{A path containing the cell ranger processed data.
Under this path, folders 'filtered_feature_bc_matrix' and 'raw_feature_bc_matrix' exist generally.}

\item{statPath}{A path containing the results files of step 'runScStatistics'.}

\item{savePath}{A path to save the results files. If NULL, the 'statPath' will be used instead.}

\item{bool.filter.cell}{A logical value indicating whether to filter the cells
according to the QC of 'scStatistics'.}

\item{bool.filter.gene}{A logical value indicating whether to filter the genes
according to the QC of 'scStatistics'.}

\item{anno.filter}{A vector indicating the types of genes to be filtered.
Must be some of c("mitochondrial", "ribosome", "dissociation")(default).}

\item{nCell.min}{An integer number used to filter gene. The default is 3.
Genes with the number of expressed cells less than this threshold will be filtered.}

\item{bgPercent.max}{A float number used to filter gene. The default is 0.001.
Genes with the background percentage larger than this threshold will be filtered.}

\item{hg.mm.mix}{A logical value indicating whether the sample is a mix of
human cells and mouse cells(such as PDX sample).
If TRUE, the arguments 'hg.mm.thres' and 'mix.anno' should be set to corresponding values.}
}
\value{
A list containing expr.data, cell.manifest, gene.manifest, filter.thres.
}
\description{
According to the QC results of scStatistics, filter cells and genes.
}
