% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAnnotation.R
\name{getTumorCluster}
\alias{getTumorCluster}
\title{getTumorCluster}
\usage{
getTumorCluster(cell.annotation, epi.thres = 0.6, malign.thres = 0.8)
}
\arguments{
\item{cell.annotation}{A data.frame of cells' annotation containing predicted cell typea and estimated cell malignant type.}

\item{epi.thres}{A threshold for epithelial cell percent to decide putative tumor clusters.}

\item{malign.thres}{A threshold for malignant cell percent to decide putative tumor clusters.}
}
\value{
A list of identified tumor clusters. If no clusters are found, return NULL.
}
\description{
Identify tumor clusters according to the results of cell type prediction and cell malignancy estimatation.
}
