% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtypeFunction.R
\name{HRG}
\alias{HRG}
\title{HRG}
\usage{
HRG(counts, k = 1000)
}
\arguments{
\item{k}{Number of genes to select.}

\item{expr}{Rows should be cells and the last column should be "label".}
}
\value{
'out'  selected gene list(class "character)
}
\description{
HRG
}
