% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAnnotation.R
\name{markerPlot}
\alias{markerPlot}
\title{markerPlot}
\usage{
markerPlot(
  expr.data,
  coor.df,
  coor.names = c("tSNE_1", "tSNE_2"),
  features = NULL,
  add = T,
  species = "human",
  font.size = 4,
  color = "blue"
)
}
\arguments{
\item{expr.data}{A matrix of expression (gene by cell)}

\item{coor.df}{A data.frame which contains cells' 2D coordinates.}

\item{coor.names}{A vector indicating the names of two-dimension coordinate used in visualization.}

\item{features}{A vector of genes to plot.}

\item{add}{A logical value indicating whether to present the default markers.}

\item{species}{A character string indicating what species the sample belong to.
Only "human"(default) or "mouse" are allowed.}

\item{font.size}{The size of labels.}

\item{color}{The color of point.}
}
\value{
A list of ggplot obejects for each maker genes.
}
\description{
Generate plots of interested genes' expression profile.
}
