% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtypeFunction.R
\name{MarkerScore}
\alias{MarkerScore}
\title{MarkerScore}
\usage{
MarkerScore(
  test_set,
  marker_file_path,
  cutoff = 0.3,
  database = org.Hs.eg.db,
  metacell = FALSE
)
}
\description{
MarkerScore
}
