% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liger.R
\name{Matrix.column_norm}
\alias{Matrix.column_norm}
\title{Perform fast and memory-efficient normalization operation on sparse matrix data.}
\usage{
Matrix.column_norm(A)
}
\arguments{
\item{A}{Sparse matrix DGE.}
}
\description{
Perform fast and memory-efficient normalization operation on sparse matrix data.
}
