% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtypeFunction.R
\name{MLEstimate}
\alias{MLEstimate}
\title{MLEstimate}
\usage{
MLEstimate(
  test,
  prob,
  lambda = NULL,
  weighted.markers = NULL,
  dropout.modeling = FALSE
)
}
\arguments{
\item{test}{Rows should be cells and columns should be genes.}

\item{prob}{trained scibet model  Rows should be genes
and columns should be cell types.The genes must be matched with test_r}
}
\value{
'cellType' and 'cell*label' likelihood matrix
}
\description{
MLEstimate
}
