% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liger.R
\name{normalize}
\alias{normalize}
\title{Normalize raw datasets to column sums}
\usage{
normalize(object)
}
\arguments{
\item{object}{\code{liger} object.}
}
\value{
\code{liger} object with norm.data slot set.
}
\description{
This function normalizes data to account for total gene expression across a cell.
}
