% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Harmony.R, R/liger.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe operator}
\usage{
lhs \%>\% rhs
}
\value{
return value of rhs function.

return value of rhs function.
}
\description{
Pipe operator

Pipe operator
}
\keyword{internal}
