% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAnnotation.R
\name{plotCellInteraction}
\alias{plotCellInteraction}
\title{plotCellInteraction}
\usage{
plotCellInteraction(stat.df, cell.annotation)
}
\arguments{
\item{stat.df}{A data.frame of cell sets interaction result.}

\item{cell.annotation}{A data.frame of cells' annotation containing the cells' cluster and type.}
}
\value{
A plot showing the result of cell interaction.
}
\description{
plotCellInteraction
}
