% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAnnotation.R
\name{plotGeneSet}
\alias{plotGeneSet}
\title{plotGeneSet}
\usage{
plotGeneSet(cell.annotation, prefix = "GS__", bool.limit = T, savePath = NULL)
}
\arguments{
\item{cell.annotation}{A data.frame of cells' annotation containing gene set signature scores.}

\item{prefix}{A prefix string of column names for gene sets.}

\item{bool.limit}{A logical value indicating whether to set upper and lower limit when plot heatmap.}

\item{savePath}{A path to save the results files. If NULL, the 'statPath' will be used instead.}
}
\value{
A heatmap for gene set signature scores.
}
\description{
plotGeneSet
}
