% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnvFunction.R
\name{plotMalignancy}
\alias{plotMalignancy}
\title{plotMalignancy}
\usage{
plotMalignancy(
  cell.annotation,
  coor.names = c("tSNE_1", "tSNE_2"),
  savePath = NULL
)
}
\arguments{
\item{coor.names}{A vector indicating the names of two-dimension coordinate used in visualization.}

\item{savePath}{A path to save the results files. If NULL, the 'statPath' will be used instead.}
}
\value{
A plot list.
}
\description{
plotMalignancy
}
