% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAnnotation.R
\name{pointDRPlot}
\alias{pointDRPlot}
\title{pointDRPlot}
\usage{
pointDRPlot(cell.annotation, value, coor.names = c("tSNE_1", "tSNE_2"),
  colors = NULL, discrete = T, limit.quantile = 0,
  legend.position = "right", legend.title = NULL)
}
\arguments{
\item{cell.annotation}{A data.frame of cells' annotation containing the cells' coordinates and index to be colored.}

\item{value}{The column name of cell.annotation, which is mapped to the colors of points.}

\item{coor.names}{A vector indicating the names of two-dimension coordinate used in visualization.}

\item{colors}{An array of colors used to show the gredients or type of points. If NULL, the default colors will be used.}

\item{discrete}{A logical value indicating whether the value column is discrete or not.}

\item{limit.quantile}{A quantile threshold to limit the data and reduce the influence of outliers.}

\item{legend.position}{The position of legends ("none", "left", "right", "bottom", "top", or two-element numeric vector).}

\item{legend.title}{The title of legends.}
}
\value{
A ggplot object for the scatter plot.
}
\description{
Plot scatter for cells.
}
