% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAnnotation.R
\name{predCellType}
\alias{predCellType}
\title{predCellType}
\usage{
predCellType(X.test, ct.templates = NULL)
}
\arguments{
\item{X.test}{A cells expression matrix (row for genes, column for cells).}

\item{ct.templates}{A list of gene weight vectors for each cell type.}
}
\value{
A list of predicted cell types and the relative correlations.
}
\description{
predCellType
}
