% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmeIdentification.R
\name{predMalignantCell}
\alias{predMalignantCell}
\title{predMalignantCell}
\usage{
predMalignantCell(
  expr,
  cell.annotation,
  savePath,
  coor.names = c("tSNE_1", "tSNE_2"),
  MALIGNANT.THRES = 0.5,
  model.path = NULL,
  genes.path = NULL
)
}
\arguments{
\item{expr}{A Seurat object.}

\item{cell.annotation}{A data.frame of cells' annotation.}

\item{savePath}{A path to save the results files. If NULL, the 'statPath' will be used instead.}

\item{coor.names}{A vector indicating the names of two-dimension coordinate used in visualization.}

\item{MALIGNANT.THRES}{A threshold of xgboost score
to decide whether a cell is malignant. Default is 0.5.}
}
\value{
A list of cell.annotation and malignancy plots
}
\description{
predMalignantCell
}
