% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmeIdentification.R
\name{predSubType_Scoring}
\alias{predSubType_Scoring}
\title{predSubType_Scoring (Old version, scoring)}
\usage{
predSubType_Scoring(
  expr,
  submodel.path,
  markers.path,
  savePath,
  celltype.list,
  dropout.modeling,
  unknown.cutoff,
  umap.plot = FALSE
)
}
\arguments{
\item{submodel.path}{A folder containing all models preset for cell subtype annotation(.csv file)
The default is NULL and all models prepared in this package will be used.}

\item{markers.path}{A folder containing all marker files preset for cell subtype annotation(.txt file)
The default is NULL and all marker files prepared in this package will be used.}

\item{savePath}{A path to save the results files. If NULL, the 'statPath' will be used instead.}

\item{celltype.list}{A list of cell types for subtype annotation, which depends on either rough annotation or user's input.}

\item{unknown.cutoff}{A threshold for assignment of unknown label. Default is 0.3.}

\item{test_set}{An expression matrix.
Rows should be cells and the last column should be "rough label".}
}
\value{
A list of fine.labels containing all possible celltypes
}
\description{
predSubType_Scoring (Old version, scoring)
}
