% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmeIdentification.R
\name{predSubType_XGBoost}
\alias{predSubType_XGBoost}
\title{predSubType_XGBoost}
\usage{
predSubType_XGBoost(expr, submodel.path, savePath, celltype.list, umap.plot)
}
\arguments{
\item{submodel.path}{A folder containing all models preset for cell subtype annotation(.csv file)
The default is NULL and all models prepared in this package will be used.}

\item{savePath}{A path to save the results files. If NULL, the 'statPath' will be used instead.}

\item{celltype.list}{A list of cell types for subtype annotation, which depends on either rough annotation or user's input.}

\item{test_set}{An expression matrix.
Rows should be cells and the last column should be "rough label".}
}
\value{
A list of fine.labels containing all possible celltypes
}
\description{
predSubType_XGBoost
}
