% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scStatistics.R
\name{prepareData}
\alias{prepareData}
\title{prepareData}
\usage{
prepareData(
  samplePath,
  species = "human",
  hg.mm.mix = F,
  hg.mm.thres = 0.9,
  mix.anno = c(human = "hg19", mouse = "mm10")
)
}
\arguments{
\item{samplePath}{A path containing the cell ranger processed data.}

\item{species}{A character string indicating what species the sample belong to.
Must be one of "human"(default) and "mouse".}

\item{hg.mm.mix}{A logical value indicating whether the sample is a mix of
human cells and mouse cells(such as PDX sample).
If TRUE, the arguments 'hg.mm.thres' and 'mix.anno' should be set to corresponding values.}

\item{hg.mm.thres}{A float-point threshold within [0.5, 1] to identify human and mouse cells.
Cells with UMI percentage of single species larger than the threshold are labeled human or mouse cells.
The default is 0.6.}

\item{mix.anno}{A vector to indicate the prefix of genes from different species.
The default is c("human" = "hg19", "mouse" = "mm10").}
}
\value{
A list of expr.data, cell.manifest, gene.manifest, raw.data, min.nUMI, cr.version and run.emptydrop
}
\description{
prepareData
}
