% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAnnotation.R
\name{runCellClassify}
\alias{runCellClassify}
\title{runCellClassify}
\usage{
runCellClassify(expr, cell.annotation, savePath, ct.templates = NULL)
}
\arguments{
\item{expr}{A Seurat object return by function 'runSeurat'.}

\item{cell.annotation}{A data.frame of cells' annotation return by function 'runSeurat'.}

\item{savePath}{A path to save the results files. If NULL, the 'statPath' will be used instead.}

\item{ct.templates}{A list of vectors of several cell type templates.
The default is NULL and the templates prepared in this package will be used.}
}
\value{
A list of updated Seurat object, cell.annotation, and the plots for cell type annotation.
}
\description{
Use a one-class logistic regression (OCLR) model to predict cancer microenvironment cell types.
}
