% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAnnotation.R
\name{runCellCycle}
\alias{runCellCycle}
\title{runCellCycle}
\usage{
runCellCycle(expr)
}
\arguments{
\item{expr}{A Seurat object.}
}
\value{
A array of cell cycle scores.
}
\description{
Estimate cell cycle scores.
}
