% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmeIdentification.R
\name{runCellSubtypeClassify}
\alias{runCellSubtypeClassify}
\title{runCellSubtypeClassify}
\usage{
runCellSubtypeClassify(
  expr,
  submodel.path,
  markers.path,
  savePath,
  celltype.list,
  dropout.modeling,
  unknown.cutoff,
  umap.plot
)
}
\arguments{
\item{expr}{A Seurat object.}

\item{submodel.path}{A folder containing all models preset for cell subtype annotation(.csv file)
The default is NULL and all models prepared in this package will be used.}

\item{markers.path}{A folder containing all marker files preset for cell subtype annotation(.txt file)
The default is NULL and all marker files prepared in this package will be used.}

\item{savePath}{A path to save the results files. If NULL, the 'statPath' will be used instead.}

\item{celltype.list}{A list of cell types for subtype annotation, which depends on either rough annotation or user's input.}
}
\value{
A list of fine.labels and similarity matrix
}
\description{
runCellSubtypeClassify
}
