% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAnnotation.R
\name{runDoublet}
\alias{runDoublet}
\title{runDoublet}
\usage{
runDoublet(expr, method = "cxds", pc.use = 30)
}
\arguments{
\item{expr}{A seurat object.}

\item{method}{The method to estimate doublet score. The default is "cxds".}

\item{pc.use}{An integer number indicating the number of PCs to use as input features. The default is 30.}
}
\value{
An array of doublet scores.
}
\description{
runDoublet
}
