% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAnnotation.R
\name{runGeneSets}
\alias{runGeneSets}
\title{runGeneSets}
\usage{
runGeneSets(expr, geneSets, method = "average")
}
\arguments{
\item{expr}{A Seurat object.}

\item{geneSets}{A list of gene sets to be analyzed. The default is NULL and 50 hallmark gene sets from MSigDB will be used.}

\item{method}{The method to be used in calculate gene set scores. Currently, only "average" and "GSVA" are allowed.}
}
\value{
A data.frame of calculated gene set signature scores.
}
\description{
Calculate gene set signature scores for cells.
}
