% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnvFunction.R
\name{runMalignancy}
\alias{runMalignancy}
\title{runMalignancy}
\usage{
runMalignancy(dataPath, statPath, savePath, cell.annotation, expr,
  cutoff = 0.1, minCell = 3, p.value.cutoff = 0.5,
  coor.names = c("tSNE_1", "tSNE_2"), hg.mm.mix = F)
}
\arguments{
\item{dataPath}{A path containing the cell ranger processed data.
Under this path, folders 'filtered_feature_bc_matrix' and 'raw_feature_bc_matrix' exist generally.}

\item{statPath}{A path containing the results files of step 'runScStatistics'.}

\item{savePath}{A path to save the results files. If NULL, the 'statPath' will be used instead.}

\item{cell.annotation}{A data.frame of cells' annotation}

\item{expr}{A Seurat object.}

\item{cutoff}{The cut-off for min average read counts per gene among
reference cells. The default is 0.1)}

\item{minCell}{An integer number used to filter gene. The default is 3.}

\item{p.value.cutoff}{The p-value to decide whether the distribution of
malignancy score is bimodality.}

\item{coor.names}{A vector indicating the names of two-dimension coordinate used in visualization.}

\item{hg.mm.mix}{A logical value indicating whether the sample is a mix of
human cells and mouse cells(such as PDX sample).
If TRUE, the arguments 'hg.mm.thres' and 'mix.anno' should be set to corresponding values.}
}
\value{
A list of cnvList, reference malignancy score, seurat object,
cell.annotatino, bimodal.pvalue, malign.thres, and all generated plots.
}
\description{
runMalignancy
}
