% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scCombination.R
\name{runScCombination}
\alias{runScCombination}
\title{runScCombination}
\usage{
runScCombination(
  single.savePaths,
  sampleNames,
  savePath,
  combName,
  authorName = NULL,
  comb.method = "NormalMNN",
  vars.to.regress = c("nCount_RNA", "mito.percent", "ribo.percent"),
  pc.use = 30,
  resolution = 0.8,
  clusterStashName = "comb.cluster",
  show.features = NULL,
  bool.add.features = T,
  bool.runDiffExpr = T,
  n.markers = 5,
  sample.colors = NULL,
  species = "human",
  genome = "hg19",
  hg.mm.mix = F,
  bool.runCellClassify = T,
  ct.templates = NULL,
  coor.names = c("tSNE_1", "tSNE_2"),
  bool.runMalignancy = T,
  cnv.ref.data = NULL,
  cnv.referAdjMat = NULL,
  cutoff = 0.1,
  p.value.cutoff = 0.5,
  bool.intraTumor = T,
  bool.runCellCycle = T,
  bool.runStemness = T,
  bool.runGeneSets = T,
  geneSets = NULL,
  geneSet.method = "average",
  bool.runExprProgram = T,
  nmf.rank = 50,
  genReport = T
)
}
\arguments{
\item{single.savePaths}{A vecotr of paths containing the results files of step 'runScAnnotation' for each sample.}

\item{sampleNames}{A vector of labels for all samples.}

\item{savePath}{A path to save the results files. If NULL, the 'statPath' will be used instead.}

\item{combName}{A label for the combined samples.}

\item{authorName}{A character string for authors name and will be shown in the report.}

\item{comb.method}{The method to combine samples. The default is "NormalMNN". "Harmony", "NormalMNN", "SeuratMNN", "Raw", "Regression" and "LIGER" are optional.}

\item{vars.to.regress}{A vector indicating the variables to regress out in R package Seurat.
The default is c("nCount_RNA", "mito.percent", "ribo.percent").}

\item{pc.use}{An integer number indicating the number of PCs to use as input features. The default is 30.}

\item{resolution}{A float number used in function 'FindClusters' in Seurat. The default is 0.8.}

\item{clusterStashName}{A character string used as the name of cluster identies. The default is "default".}

\item{show.features}{A list or vector for genes to be plotted in 'markerPlot'.}

\item{bool.add.features}{A logical value indicating whether to add default features to 'show.features' or not.}

\item{bool.runDiffExpr}{A logical value indicating whether to perform differential expressed analysis.}

\item{n.markers}{An integer indicating the number of differential expressed genes showed in the plot. The defalut is 5.}

\item{species}{A character string indicating what species the sample belong to.
Only "human"(default) or "mouse" are allowed.}

\item{genome}{A character string indicating the version of the reference gene annotation information.
This information is mainly used to infer CNV profile and estimate malignancy.
Only 'hg19' (defalut) or 'hg38' are allowed for "human" species, and only "mm10" is allowed for "mouse" species.}

\item{hg.mm.mix}{A logical value indicating whether the sample is a mix of
human cells and mouse cells(such as PDX sample).
If TRUE, the arguments 'hg.mm.thres' and 'mix.anno' should be set to corresponding values.}

\item{bool.runCellClassify}{A logical value indicating whether to predict the usual cell type. The default is TRUE.}

\item{ct.templates}{A list of vectors of several cell type templates.
The default is NULL and the templates prepared in this package will be used.}

\item{coor.names}{A vector indicating the names of two-dimension coordinate used in visualization.}

\item{bool.runMalignancy}{A logical value indicating whether to estimate malignancy.}

\item{cnv.ref.data}{An expression matrix of gene by cell, which is used as the normal reference during estimating malignancy.
The default is NULL, and an immune cells or bone marrow cells expression matrix will be used for human or mouse species, respectively.}

\item{cnv.referAdjMat}{An adjacent matrix for the normal reference data.
The larger the value, the closer the cell pair is.
The default is NULL, and a SNN matrix of the default ref.data will be used.}

\item{cutoff}{A threshold used in the CNV inference.}

\item{p.value.cutoff}{A threshold to decide weather the bimodality distribution of malignancy score is significant.}

\item{bool.intraTumor}{A logical value indicating whether to use the identified tumor clusters to
perform following intra-tumor heterogeneity analyses.}

\item{bool.runCellCycle}{A logical value indicating whether to estimate cell cycle scores.}

\item{bool.runStemness}{A logical value indicating whether to estimate stemness scores.}

\item{bool.runGeneSets}{A logical value indicating whether to estimate gene sets signature scores.}

\item{geneSets}{A list of gene sets to be analyzed. The default is NULL and 50 hallmark gene sets from MSigDB will be used.}

\item{geneSet.method}{The method to be used in calculate gene set scores. Currently, only "average" and "GSVA" are allowed.}

\item{bool.runExprProgram}{A logical value indicating whether to run non-negative matrix factorization (NMF) to identify expression programs.}

\item{nmf.rank}{An integer of decomposition rank used in NMF.}

\item{genReport}{A logical value indicating whether to generate a .html/.md report (suggest to set TRUE).}
}
\value{
A results list with all useful objects used in the function.
}
\description{
Perform multi-samples analyses.
}
