% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scStatistics.R
\name{runScStatistics}
\alias{runScStatistics}
\title{runScStatistics}
\usage{
runScStatistics(
  dataPath,
  savePath,
  authorName = NULL,
  sampleName = "sc",
  species = "human",
  hg.mm.mix = F,
  hg.mm.thres = 0.6,
  mix.anno = c(human = "hg19", mouse = "mm10"),
  bg.spec.genes = NULL,
  bool.runSoupx = F,
  genReport = T
)
}
\arguments{
\item{dataPath}{A path containing the cell ranger processed data.
Under this path, folders 'filtered_feature_bc_matrix' and 'raw_feature_bc_matrix' exist generally.}

\item{savePath}{A path to save the results files(suggest to create a foler named by sample name).}

\item{authorName}{A character string for authors name and will be shown in the report.}

\item{sampleName}{A character string giving a label for this sample.}

\item{species}{A character string indicating what species the sample belong to.
Must be one of "human"(default) and "mouse".}

\item{hg.mm.mix}{A logical value indicating whether the sample is a mix of
human cells and mouse cells(such as PDX sample).
If TRUE, the arguments 'hg.mm.thres' and 'mix.anno' should be set to corresponding values.}

\item{hg.mm.thres}{A float-point threshold within [0.5, 1] to identify human and mouse cells.
Cells with UMI percentage of single species larger than the threshold are labeled human or mouse cells.
The default is 0.6.}

\item{mix.anno}{A vector to indicate the prefix of genes from different species.
The default is c("human" = "hg19", "mouse" = "mm10").}

\item{bg.spec.genes}{A list of backgroud specific genes, which are used to remove ambient genes' influence.}

\item{genReport}{A logical value indicating whether to generate a .html/.md report (suggest to set TRUE).}
}
\value{
A results list with all useful objects used in the function.
}
\description{
perform cell QC, gene QC, visualization and give suggested thresholds.
}
