% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liger.R
\name{runWilcoxon}
\alias{runWilcoxon}
\title{Perform Wilcoxon rank-sum test}
\usage{
runWilcoxon(object, data.use = "all", compare.method)
}
\arguments{
\item{object}{\code{liger} object.}

\item{data.use}{This selects which dataset(s) to use. (default 'all')}

\item{compare.method}{This indicates the metric of the test. Either 'clusters' or 'datasets'.}
}
\value{
A 10-columns data.frame with test results.
}
\description{
Perform Wilcoxon rank-sum tests on specified dataset using given method.
}
