% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scibet_visualization}
\alias{scibet_visualization}
\title{Construct a convenient Seurat pipeline}
\usage{
Scibet_visualization(dataset, label)
}
\arguments{
\item{dataset}{The expression dataframe, 
with rows being cells, and columns being genes. 
The last column should be "label".}

\item{label}{groundtruth or output of scibet function Test.}
}
\description{
Construct a convenient Seurat pipeline
}
