% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtypeFunction.R
\name{SelectCells}
\alias{SelectCells}
\title{SelectCells}
\usage{
SelectCells(object, marker_file_path, cutoff = 0.75, database = org.Hs.eg.db)
}
\arguments{
\item{object}{Seurat object of training set.}
}
\value{
representative.index A user-defined set of cells for training.
}
\description{
SelectCells
}
