% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtypeFunction.R
\name{SelectGenes}
\alias{SelectGenes}
\title{SelectGenes}
\usage{
SelectGenes(expr, method = "Entropy", k = 2000)
}
\arguments{
\item{expr}{Rows should be cells and the last column should be "label".}
}
\value{
geneset A user-defined set of genes for prediction.
}
\description{
SelectGenes
}
