% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmeIdentification.R
\name{similarityCalculation}
\alias{similarityCalculation}
\title{similarityCalculation}
\usage{
similarityCalculation(fine.labels, savePath)
}
\arguments{
\item{fine.labels}{annotation results (barcode-subtype) of function"predSubType"}

\item{savePath}{A path to save the results files. If NULL, the 'statPath' will be used instead.}
}
\value{
similarity matrixes of all possible celltypes
}
\description{
similarityCalculation
}
