% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{SimilarityMap}
\alias{SimilarityMap}
\title{Correlation heatmap and hierarchical clustering}
\usage{
SimilarityMap(
  plot.title,
  reference,
  similarity.mar,
  similarity.var = NULL,
  number.digits = 2,
  number.cex = 1,
  tl.cex = 1
)
}
\description{
Correlation heatmap and hierarchical clustering
}
