% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scibetFunction.R
\name{Test}
\alias{Test}
\title{Predict cell type and show accuracy}
\usage{
Test(prob, test, result="list")
}
\arguments{
\item{prob}{A matrix generated by Train(cellType-genes prob-matrix)}

\item{test}{testset with rows being cells, and columns being genes.}
}
\value{
cellType prediction result
}
\description{
Predict cell type and show accuracy
}
