% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scibetFunction.R
\name{Train}
\alias{Train}
\title{Train SciBet model and return cellType-genes prob-matrix}
\usage{
Train(expr, geneset, k=1000)
}
\arguments{
\item{expr}{The reference expression dataframeThe expression dataframe, with rows being cells, and columns being genes. The last column should be "label".}

\item{geneset}{A user-defined set of genes for prediction.}
}
\value{
reference model(cellType-genes prob-matrix)
}
\description{
Train SciBet model and return cellType-genes prob-matrix
}
