% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAnnotation.R
\name{runCellClassify}
\alias{runCellClassify}
\title{runCellClassify}
\usage{
runCellClassify(
  expr,
  cell.annotation,
  coor.names = c("tSNE_1", "tSNE_2"),
  savePath,
  ct.templates = NULL,
  species = "human"
)
}
\arguments{
\item{expr}{A Seurat object.}

\item{cell.annotation}{A data.frame of cells' annotation.}

\item{coor.names}{A vector indicating the names of two-dimension coordinate used in visualization.}

\item{savePath}{A path to save the results files. If NULL, the 'statPath' will be used instead.}

\item{ct.templates}{A list of vectors of several cell type templates.
The default is NULL and the templates prepared in this package will be used.}

\item{species}{A character string indicating what species the sample belong to.
Only "human"(default) or "mouse" are allowed.}
}
\value{
A list of updated Seurat object, cell.annotation, and the plots for cell type annotation.
}
\description{
Use a one-class logistic regression (OCLR) model to predict cancer microenvironment cell types.
}
