% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAnnotation.R
\name{runCellCycle}
\alias{runCellCycle}
\title{runCellCycle}
\usage{
runCellCycle(expr, species = "human")
}
\arguments{
\item{expr}{A Seurat object.}

\item{species}{A character string indicating what species the sample belong to.
Only "human"(default) or "mouse" are allowed.}
}
\value{
An array of cell cycle scores.
}
\description{
Estimate cell cycle scores.
}
