% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnvFunction.R
\name{runMalignancy}
\alias{runMalignancy}
\title{runMalignancy}
\usage{
runMalignancy(
  expr,
  gene.manifest,
  cell.annotation,
  savePath,
  cutoff = 0.1,
  minCell = 3,
  p.value.cutoff = 0.5,
  coor.names = c("tSNE_1", "tSNE_2"),
  ref.data = NULL,
  referAdjMat = NULL,
  species = "human",
  genome = "hg19",
  hg.mm.mix = F
)
}
\arguments{
\item{expr}{A Seurat object.}

\item{gene.manifest}{A data.frame of genes' manifest.}

\item{cell.annotation}{A data.frame of cells' annotation.}

\item{savePath}{A path to save the results files. If NULL, the 'statPath' will be used instead.}

\item{cutoff}{The cut-off for min average read counts per gene among
reference cells. The default is 0.1.}

\item{minCell}{An integer number used to filter gene. The default is 3.}

\item{p.value.cutoff}{The p-value to decide whether the distribution of
malignancy score is bimodality.}

\item{coor.names}{A vector indicating the names of two-dimension coordinate used in visualization.}

\item{ref.data}{An expression matrix of gene by cell, which is used as the normal reference.
The default is NULL, and an immune cells or bone marrow cells expression matrix will be used for human or mouse species, respectively.}

\item{referAdjMat}{An adjacent matrix for the normal reference data.
The larger the value, the closer the cell pair is.
The default is NULL, and a SNN matrix of the default ref.data will be used.}

\item{species}{A character string indicating what species the sample belong to.
Only "human"(default) or "mouse" are allowed.}

\item{genome}{A character string indicating the version of the reference gene annotation information.
This information is mainly used to infer CNV profile and estimate malignancy.
Only 'hg19' (defalut) or 'hg38' are allowed for "human" species, and only "mm10" is allowed for "mouse" species.}

\item{hg.mm.mix}{A logical value indicating whether the sample is a mix of
human cells and mouse cells(such as PDX sample).
If TRUE, the arguments 'hg.mm.thres' and 'mix.anno' should be set to corresponding values.}
}
\value{
A list of cnvList, reference malignancy score, seurat object,
cell.annotatino, bimodal.pvalue, malign.thres, and all generated plots.
}
\description{
runMalignancy
}
