% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scAnnotation.R
\name{runSeurat}
\alias{runSeurat}
\title{runSeurat}
\usage{
runSeurat(
  expr,
  savePath,
  pc.use = 30,
  resolution = 0.8,
  clusterStashName = "default",
  bool.runDiffExpr = T,
  comb.method = NULL
)
}
\arguments{
\item{expr}{A Seurat object return by prepareSeurat.}

\item{savePath}{A path to save the results files. If NULL, the 'statPath' will be used instead.}

\item{pc.use}{An integer number indicating the number of PCs to use as input features. The default is 30.}

\item{resolution}{A float number used in function 'FindClusters' in Seurat. The default is 0.8.}

\item{clusterStashName}{A character string used as the name of cluster identies. The default is "default".}

\item{bool.runDiffExpr}{A logical value indicating whether to perform differential expressed analysis.}

\item{comb.method}{The method used in combining samples. It worked only for multi-sample analysis.}
}
\value{
A list containing a Seurat object, differential expressed genes and annotation information for cells.
}
\description{
Perform usual Seurat step and cell type prediction.
}
