% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{runSurvival}
\alias{runSurvival}
\title{runSurvival}
\usage{
runSurvival(
  features,
  data,
  surv.time,
  surv.event,
  cut.off = 0.5,
  savePath = NULL
)
}
\arguments{
\item{features}{The names of marker genes or signatures to be analyzed.}

\item{data}{The data used to perform survival analysis.
It should be an expression or signature matrix with gene or signature by patient.
The row names are the features' anmes. The columns are patients' labels.}

\item{surv.time}{The survival time of patients. It should be in accord with the columns of data.}

\item{surv.event}{The status indicator of patients. 0=alive, 1=dead. It should be in accord with the columns of data.}

\item{savePath}{}
}
\value{

}
\description{
According to the marker genes or signatures expression high/low levels,
patient are divided into two groups and then survival analysis is performed.
The survival curves can be plotted.
}
